Implement CLI commands

+++ src/main/java/ukitinu/markovwords/repo/FileRepo.java
@@ -1,5 +1,6 @@
 package ukitinu.markovwords.repo;
 
+import ukitinu.markovwords.lib.Couple;
 import ukitinu.markovwords.lib.FsUtils;
 import ukitinu.markovwords.lib.Logger;
 import ukitinu.markovwords.models.Dict;
@@ -11,6 +12,7 @@ import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.HashMap;
 import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
@@ -21,7 +23,7 @@ public final class FileRepo implements Repo {
     private final Path dataPath;
     private final DataConverter dataConverter;
 
-    static FileRepo create(String dataPath) {
+    public static FileRepo create(String dataPath) {
         Path path = Path.of(dataPath);
         if (!Files.exists(path)) throw new IllegalArgumentException(path + " does not exist");
         if (!Files.isDirectory(path)) throw new IllegalArgumentException(path + " is not a directory");
@@ -44,19 +46,31 @@ public final class FileRepo implements Repo {
      * @see FilePaths to see how deleted directories are marked.
      */
     @Override
-    public Collection<String> listAll() {
+    public Couple<Collection<String>> listAll() {
         try (var files = Files.list(dataPath)) {
-            return files
+            Collection<String> names = files
                     .filter(Files::isDirectory)
                     .map(Path::getFileName)
                     .map(Path::toString)
-                    .filter(s -> !FilePaths.isDeleted(s))
                     .filter(s -> !FilePaths.isTemp(s))
-                    .collect(Collectors.toList());
+                    .toList();
+            Collection<String> visible = names.stream().filter(s -> !FilePaths.isDeleted(s)).toList();
+            Collection<String> deleted = names.stream().filter(FilePaths::isDeleted).toList();
+            return new Couple<>(visible, deleted);
         } catch (Exception e) {
             LOG.error("Unable to read dir {}: {}", dataPath.toString(), e.toString());
             throw new DataException("Unable to read data dir", e);
         }
+    }
+
+    /**
+     * Checks whether there is a dictionary with the given name.
+     */
+    @Override
+    public boolean exists(String name) {
+        Path dirPath = FilePaths.getDictDir(dataPath, name);
+        Path filePath = FilePaths.getDictFile(dataPath, name);
+        return Files.exists(dirPath) && Files.exists(filePath);
     }
 
     /**
@@ -65,11 +79,6 @@ public final class FileRepo implements Repo {
     @Override
     public Dict get(String name) {
         try {
-            if (FilePaths.isDeleted(name)) {
-                LOG.info("Cannot get deleted dict {}", name);
-                throw new DataException("Cannot get deleted dict " + name);
-            }
-
             Path filePath = FilePaths.getDictFile(dataPath, name);
             String content = FsUtils.readFile(filePath);
             return dataConverter.deserialiseDict(content);
@@ -105,11 +114,21 @@ public final class FileRepo implements Repo {
     }
 
     /**
-     * The dict's directory is renamed to mark is as deleted.
+     * If {@param permanent} is false, the dict's directory is renamed to mark is as deleted.
+     * If {@param permanent} is true, the directory is deleted permanently.
      */
     @Override
-    public void delete(String name) {
+    public void delete(String name, boolean permanent) {
         try {
+            get(name);
+
+            if (permanent) {
+                FsUtils.rmDir(FilePaths.getDictDir(dataPath, name));
+                return;
+            }
+
+            if (FilePaths.isDeleted(name)) throw new DataException("Dict " + name + " is already in deleted state");
+
             Path dirPath = FilePaths.getDictDir(dataPath, name);
             Path deletedPath = FilePaths.getDeletedDictDir(dataPath, name);
             Files.move(dirPath, deletedPath);
@@ -120,12 +139,82 @@ public final class FileRepo implements Repo {
     }
 
     /**
+     * Restores the given dictionary's directory (if deleted).
+     */
+    @Override
+    public void restore(String name) {
+        try {
+            if (!FilePaths.isDeleted(name)) throw new DataException("Dict " + name + " is not in deleted state");
+
+            Path dirPath = FilePaths.getDictDir(dataPath, name);
+            Path restoredPath = FilePaths.getRestoredDictDir(dataPath, name);
+            Files.move(dirPath, restoredPath);
+        } catch (IOException e) {
+            LOG.error("Unable to restore dict {}: {}", name, e.toString());
+            throw new DataException("Unable to restore dict " + name, e);
+        }
+    }
+
+    /**
+     * Checks whether the dictionary with the given name has grams of the given length.
+     *
+     * @param name name of the dictionary to check
+     * @param len  length of the grams.
+     * @return {@code true} if there is at least one {@code len}-gram, {@code false} otherwise.
+     */
+    @Override
+    public boolean hasGramMap(String name, int len) {
+        try {
+            if (len < 1) throw new IllegalArgumentException("gram length must be positive");
+
+            Path gramDir = FilePaths.getGramDir(dataPath, name, len);
+            if (Files.isDirectory(gramDir)) {
+                try (var entries = Files.list(gramDir)) {
+                    return entries.findFirst().isPresent();
+                }
+            }
+            return false;
+        } catch (IOException e) {
+            LOG.error("Unable to check {}-grams of {}: {}", len, name, e.toString());
+            throw new DataException("Unable to check " + len + "-grams of " + name, e);
+        }
+    }
+
+    /**
+     * Reads all the gram files of dictionary {@param name}, converts them to {@link Gram}
+     * and then returns a map containing them, with their value as key.
+     *
+     * @param name dict's name.
+     * @return gram.value-gram map of the given dictionary.
+     * @see #getGramMap(String, int) allows to get grams of given length.
+     */
+    @Override
+    public Map<String, Gram> getGramMap(String name) {
+        Path dictDir = FilePaths.getDictDir(dataPath, name);
+        try (var files = Files.list(dictDir)) {
+            Collection<Integer> lengths = files
+                    .filter(Files::isDirectory)
+                    .filter(FilePaths::isGramDir)
+                    .map(FilePaths::getGramLength)
+                    .toList();
+            Map<String, Gram> gramMap = new HashMap<>();
+            for (int len : lengths) gramMap.putAll(getGramMap(name, len));
+
+            return gramMap;
+        } catch (IOException e) {
+            LOG.error("Unable to get grams of {}: {}", name, e.toString());
+            throw new DataException("Unable to get grams of " + name, e);
+        }
+    }
+
+    /**
      * Reads all the gram files of length {@param len} of dictionary {@param name}, converts them to {@link Gram}
      * and then returns a map containing them, with their value as key.
      *
      * @param name dict's name.
      * @param len  gram length.
      * @return gram.value-gram map of the given dictionary.
+     * @see #getGramMap(String) allows to get all grams.
      */
     @Override
     public Map<String, Gram> getGramMap(String name, int len) {
@@ -160,6 +249,8 @@ public final class FileRepo implements Repo {
         upsertGramMap(gramMap, dict.name());
         replaceOriginalWithTemp(dict);
     }
+
+    //region private
 
     /**
      * Creates a temporary directory for the given dictionary, copying, if present, the current content.
@@ -282,5 +373,6 @@ public final class FileRepo implements Repo {
                     .toList();
         }
     }
+    //endregion
 
 }
